backend/requirements.txt
# AEGIS Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
python-dotenv==1.0.1
google-generativeai==0.8.0
httpx==0.28.0

# Dev dependencies
pytest==8.3.0
pytest-asyncio==0.24.0
